% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annmax_hydro.R
\name{annmax_plot}
\alias{annmax_plot}
\title{Annual/Montly maxima graph}
\usage{
annmax_plot(data_day, break_day, yea_ama_1, yea_ama_2, end_yea_ama,
  sta_yea_ama, month_sel_1, month_sel_2, stat_name, ama_var)
}
\arguments{
\item{data_day}{Matrix with values ordered by day (see function ord_day()).}

\item{break_day}{Define start year (e.g. 274 is 1.October is start hydrological year in Switzerland).}

\item{yea_ama_1}{Start year selected time window.}

\item{yea_ama_2}{End year selected time window.}

\item{month_sel_1}{Start month to use for calculating maxima.}

\item{month_sel_2}{End month to use for calculating maxima.}

\item{stat_name}{Plot title (e.g. station name)}

\item{ama_var}{Select plot type: 1) Day of the year of annual maxima (timing), 2) Magnitudes of annual maxima or 3) Montly maxima.}

\item{sta_yea_cla}{Start year of entire time series}

\item{end_yea_cla}{End year of entire time series.}
}
\description{
Plot option 'Annual Max' from shiny app melTim.
}
