% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day_ana.R
\name{day_ana}
\alias{day_ana}
\title{Analysis on daily level}
\usage{
day_ana(data, date, start_year, end_year, do_ma = F, window_width = 30,
  method_ana, break_day = 0)
}
\arguments{
\item{data}{Vector of input data}

\item{date}{Time vector}

\item{start_year}{Selected start year}

\item{end_year}{Seleced end year}

\item{do_ma}{Apply moving average filter before analysis (FALSE / TRUE)}

\item{window_width}{If do_ma is TRUE, then moving average filter with here selected window width applied}

\item{method_ana}{Selected analytical method('mean', 'median', 'sum', 'sens_slope')}

\item{break_day}{Define start day (e.g. 274 is 1.October, which is the start of  hydrological year in Switzerland).}
}
\description{
Apply functions on daily level. First, time series re-structured according to day of the year, then selected analytical function applied.
}
