% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/down_points.R
\name{down_points}
\alias{down_points}
\title{Grid around spatial point.}
\usage{
down_points(lon_lat_point_in)
}
\arguments{
\item{lon_lat_point_in}{Lat-Lon input}
}
\description{
Add points as 1 km grid around spatial point. One point turn into a spatial grind with 1 km  resolution. In total 25 spatial points returned.
}
