% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elev_buff.R
\name{elev_buff}
\alias{elev_buff}
\title{Buffer mean value}
\usage{
elev_buff(point_index, radius = 2500, points_in = grid_points,
  dem_in = dem)
}
\arguments{
\item{point_index}{Index which point of Spatial points (points_in) used}

\item{radius}{Length side of square buffer}

\item{points_in}{Spatial points}

\item{dem_in}{DEM}
}
\description{
Get mean elevation of squared buffer around grid point.
}
