% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_coords.R
\name{extract_coords}
\alias{extract_coords}
\title{Extract coordinates from Polygon}
\usage{
extract_coords(sp_df)
}
\arguments{
\item{sp_df}{SpatialPolyognsDataFrame}
}
\description{
Extract coordinates from SpatialPolyognsDataFrame
}
