% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase_lags.R
\name{f_phase_lag}
\alias{f_phase_lag}
\title{Phase lags}
\usage{
f_phase_lag(grdc_data, break_day, end_yea_cla, sta_yea_cla)
}
\arguments{
\item{break_day}{Define start year (e.g. 274 is 1.October is start hydrological year in Switzerland).}

\item{end_yea_cla}{End year of entire time series.}

\item{sta_yea_cla}{Start year of entire time series}

\item{data_day}{Matrix with values ordered by day (see function ord_day()).}
}
\description{
Calculations of phase lags between mean hydrograph and individual years using package: devtools::install_github("laubblatt/phaselag").
}
