% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cube_index.R
\name{get_cube_index_row}
\alias{get_cube_index_row}
\title{Cube index from points inside polygon.}
\usage{
get_cube_index_row(val_in, lons_in = lon2D, col_or_row = "row")
}
\arguments{
\item{val_in}{Longitude values}

\item{lons_in}{Longitude values of data cube points (matrix)}
}
\description{
Get row number of matching point.
}
