% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_scale.R
\name{image_scale}
\alias{image_scale}
\title{Color scale for image plot.}
\usage{
image_scale(z, zlim, col, breaks, horiz = TRUE, ylim = NULL,
  xlim = NULL, ...)
}
\arguments{
\item{z}{Input data matrix. Results of moving average analysis.}

\item{zlim}{Maximum and minimum values.}

\item{col}{Vector of colors for z-value.}

\item{breaks}{Color breaks in z-range.}

\item{horiz}{Orientation of image scale}

\item{ylim}{...}

\item{xlim}{...}
}
\value{
Scale for image plot.
}
\description{
Plot color scale of image plot.
}
