% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin_trend.R
\name{lin_trend}
\alias{lin_trend}
\title{Linear trend data}
\usage{
lin_trend(data_in, index_in)
}
\arguments{
\item{data_in}{Numeric data vector.}

\item{index_in}{Index 1 to length of data vector.}
}
\description{
Calculate linear trend using function glm for fitting generalized linear models.
}
