% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loess_na_restore.R
\name{loess_NA_restore}
\alias{loess_NA_restore}
\title{Local Polynomial Regression Fitting.}
\usage{
loess_NA_restore(data_in, smoo_val = 0.2, NA_restore = TRUE)
}
\arguments{
\item{data_in}{Input vector to be smoothed}

\item{smoo_val}{Smoothing value.}

\item{NA_restore}{Restore NAs yes or no (T or F)}
}
\value{
Smoothed vector.
}
\description{
Local Polynomial Regression Fitting with NA-restoring.
}
