% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functs_na.R
\name{max_na}
\alias{max_na}
\title{Maximum with NA values.}
\usage{
max_na(x)
}
\arguments{
\item{x}{numeric or character arguments; see also R documentation for \link{min}.}
}
\value{
The minimum value in \code{x}.
}
\description{
Function to determine maximum value automatically handling NAs. Calculates
maximum value disregarding NAs, except when all values are NA.
}
\examples{
max_na(c(2 ,3, NA))
max_na(c(NA, NA, NA))
}
