% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functs_na.R
\name{mea_na}
\alias{mea_na}
\title{Arithmetic mean with NA values.}
\usage{
mea_na(x)
}
\arguments{
\item{x}{see arguments in R documentation for \link{mean}.}
}
\value{
The arithmetic mean of values in \code{x}.
}
\description{
Function for arithmetic mean automatically handling NAs. Calculates arithmetic mean value
disregarding NAs, except when all values are NA.
}
\examples{
mea_na(c(2 ,3, NA))
mea_na(c(NA, NA, NA))
}
