% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functs_na.R
\name{mea_na_thres}
\alias{mea_na_thres}
\title{Arithmetic mean with percentage threshold for NA values.}
\usage{
mea_na_thres(x, na_thres = 0.5)
}
\arguments{
\item{x}{see arguments in R documentation for \link{mean}.}

\item{na_thres}{threshold value for percentage NAs in x; between 0 and 1}
}
\value{
The arithmetic mean of values in \code{x}.
}
\description{
Function for arithmetic mean automatically handling NAs. Calculates arithmetic mean value
only when percentage of NAs is below given threshold.
}
\examples{
mea_na_thres(c(2 ,3, 4, NA), na_thres=0.5)
mea_na_thres(c(2 ,3, 4, NA), na_thres=0.1)
}
