% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_hydro.R
\name{mean_hydro}
\alias{mean_hydro}
\title{Mean graph}
\usage{
mean_hydro(
  data_day,
  break_day,
  yea_cla_1,
  yea_cla_2,
  yea_cla_3,
  yea_cla_4,
  end_yea_cla,
  sta_yea_cla,
  stat_name
)
}
\arguments{
\item{data_day}{Matrix with values ordered by day (see function \code{\link{ord_day}}).}

\item{break_day}{Define start day (e.g. 274 is 1.October, is start of the hydrological year in Switzerland).}

\item{yea_cla_1}{Start year of time window 1.}

\item{yea_cla_2}{End year of time window 1.}

\item{yea_cla_3}{Start year of time window 2.}

\item{yea_cla_4}{End year of time window 2.}

\item{end_yea_cla}{End year of entire time series.}

\item{sta_yea_cla}{Start year of entire time series}

\item{stat_name}{Plot title (e.g. station name)}
}
\description{
Plot option 'Mean graph' from shiny app 'Hydro Explorer'.
}
