% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functs_na.R
\name{med_na}
\alias{med_na}
\title{Median with NA values.}
\usage{
med_na(x)
}
\arguments{
\item{x}{see R documentation for \link{median}.}
}
\value{
The median values in \code{x}.
}
\description{
Function to determine median value automatically handling NAs. Calculates
median value disregarding NAs, except when all values are NA.
}
\examples{
med_na(c(2 ,3, NA))
med_na(c(NA, NA, NA))
}
