% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functs_na.R
\name{min_na}
\alias{min_na}
\title{Minimum with NA values.}
\usage{
min_na(x)
}
\arguments{
\item{x}{numeric or character arguments; see also R documentation for \link{min}.}
}
\value{
The minimum value in \code{x}.
}
\description{
Function to determine minimum value automatically handling NAs. Calculates
minimum value disregarding NAs, except when all values are NA.
}
\examples{
min_na(c(2 ,3, NA))
min_na(c(NA, NA, NA))
}
