% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_check.R
\name{na_check}
\alias{na_check}
\title{Check for NA}
\usage{
na_check(data_in)
}
\arguments{
\item{file_path}{Path data table.}
}
\description{
Check vector if NAs. Only use arrays with no NAs. If one NA all put to NA.
}
