% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perc_hydro.R
\name{perc_hydro}
\alias{perc_hydro}
\title{Percentile graph}
\usage{
perc_hydro(
  data_day,
  yea_per_1,
  yea_per_2,
  yea_per_3,
  yea_per_4,
  end_yea_per,
  sta_yea_per,
  stat_name,
  plot_per,
  perc_sel
)
}
\arguments{
\item{data_day}{Matrix with values ordered by day (see function \code{\link{ord_day}}).}

\item{yea_per_1}{Start year of time window 1.}

\item{yea_per_2}{End year of time window 1.}

\item{yea_per_3}{Start year of time window 2.}

\item{yea_per_4}{End year of time window 2.}

\item{end_yea_per}{End year of entire time series.}

\item{sta_yea_per}{Start year of entire time series}

\item{stat_name}{Plot title (e.g. station name)}

\item{plot_per}{Plot type: Line plot showing one selected probability level (see parameter 'perc_sel') or image plot showing difference for all probability levels between the two time windows.}

\item{perc_sel}{Probablility level displayed (for option 'Line Plot' parameter 'plot_per')}
}
\description{
Plot option 'Percentile graph' from shiny app 'Hydro Explorer'.
}
