% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_hydro.R
\name{raster_hydro}
\alias{raster_hydro}
\title{Raster graph}
\usage{
raster_hydro(data_day, break_day, sta_yea_cla, end_yea_cla, stat_name)
}
\arguments{
\item{data_day}{Matrix with values ordered by day (see function ord_day()).}

\item{break_day}{Define start year (e.g. 274 is 1.October is start hydrological year in Switzerland).}

\item{sta_yea_cla}{Start year of time series visualized.}

\item{end_yea_cla}{End year of time series visualized.}

\item{stat_name}{Plot title (e.g. station name)}
}
\description{
Plot option 'Raster graph' from shiny app melTim.
}
