% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_camels.R
\name{read_camels}
\alias{read_camels}
\title{Read CAMELS data from file.}
\usage{
read_camels(file_path)
}
\arguments{
\item{file_path}{Path data time series data.}
}
\description{
Read in CAMELS observed streamflow from USGS (https://ncar.github.io/hydrology/datasets/CAMELS_timeseries).
}
