% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_camels_br.R
\name{read_camels_br}
\alias{read_camels_br}
\title{Read CAMELS-BR data from file.}
\usage{
read_camels_br(file_path)
}
\arguments{
\item{file_path}{Path data time series data.}
}
\description{
Read in CAMELS-BR observed streamflow (https://zenodo.org/record/3964745).
}
