% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_camels_gb.R
\name{read_camels_gb}
\alias{read_camels_gb}
\title{Read CAMELS-GB data from file.}
\usage{
read_camels_gb(file_path)
}
\arguments{
\item{file_path}{Path data time series data.}
}
\description{
Read in CAMELS-GB observed streamflow (https://doi.org/10.5285/8344e4f3-d2ea-44f5-8afa-86d2987543a9).
}
