% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_grdc.R
\name{read_grdc}
\alias{read_grdc}
\title{Read GRDC data from file.}
\usage{
read_grdc(file_path)
}
\arguments{
\item{file_path}{Path data table.}
}
\description{
Read in GRDC data table from repository.
}
