% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_lamah.R
\name{read_lamah}
\alias{read_lamah}
\title{Read LamaH data from file.}
\usage{
read_lamah(file_path)
}
\arguments{
\item{file_path}{Path data time series data.}
}
\description{
Read in LamaH data time series (https://doi.org/10.5281/zenodo.452524) from file.
}
