% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sens_slo.R
\name{sens_slo}
\alias{sens_slo}
\title{Sen´s slope}
\usage{
sens_slo(data_in, cover_thresh = 0.9)
}
\arguments{
\item{data_in}{Vector of input data}

\item{Trend}{only calculated when percentage of NAs in input vecotr does not exceed cover threshold (0-1)}
}
\description{
Calclation linear trend using Sen´s slope.
}
