% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functs_na.R
\name{sum_na}
\alias{sum_na}
\title{Sum with NA values.}
\usage{
sum_na(x)
}
\arguments{
\item{x}{numeric or complex or logical vectors; see R documentation for \link{sum}.}
}
\value{
The sum of \code{x}.
}
\description{
Function to determine sum automatically handling NAs. Calculates
sum disregarding NAs, except when all values are NA.
}
\examples{
sum_na(c(2 ,3, NA))
sum_na(c(NA, NA, NA))
}
