% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volu_time.R
\name{volu_time}
\alias{volu_time}
\title{Volume timing graph}
\usage{
volu_time(
  day_cross,
  sta_yea_cla,
  end_yea_cla,
  break_day,
  day_cross_slo,
  day_cross_mea,
  day_cross_day,
  stat_name
)
}
\arguments{
\item{day_cross}{Matrix with DOYs when certain fraction of annual runoff was recorded (currently 25, 50 and 75 \%).}

\item{sta_yea_cla}{Start year of time series investigated.}

\item{end_yea_cla}{Endyear of time series investigated.}

\item{break_day}{Define start year (e.g. 274 is 1.October is start hydrological year in Switzerland).}

\item{day_cross_slo}{Trend in DOYs for fractions 25/50/75 \% of annual runoff in days/dec}

\item{day_cross_mea}{Mean DOY for fractions 25/50/75 \% of annual runoff.}

\item{day_cross_day}{Trend in DOYs for fractions 25/50/75 \% of annual runoff for the entire time frame.}

\item{stat_name}{Plot title (e.g. station name)}
}
\description{
Plot option 'Volume timing' from shiny app melTim.
}
