% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elev_buff.R
\name{elev_buff}
\alias{elev_buff}
\title{Buffer mean value}
\usage{
elev_buff(point_in, radius = 2500, dem_in)
}
\arguments{
\item{point_in}{Spatial point}

\item{radius}{Set length side of square buffer by defining radius, length point to side, is half total square size}

\item{dem_in}{DEM}
}
\description{
Get mean elevation of squared buffer around grid point.
}
